[Refactor Name] code -> join
