+++ app/src/main/java/com/coeater/android/join/JoinActivity.kt
@@ -1,13 +1,10 @@
-package com.coeater.android.code
+package com.coeater.android.join
 
 import android.Manifest
-import android.app.PendingIntent.getActivity
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import android.view.WindowManager
-import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
@@ -15,36 +12,32 @@ import androidx.lifecycle.ViewModelProviders
 import com.coeater.android.R
 import com.coeater.android.api.UserManageProvider
 import com.coeater.android.api.provideMatchApi
-import com.coeater.android.api.provideUserApi
-import com.coeater.android.main.MainActivity
-import com.coeater.android.main.fragment.OneOnOneCodeFragment
 import com.coeater.android.model.RoomResponse
 import com.coeater.android.webrtc.CallActivity
 import com.gun0912.tedpermission.PermissionListener
 import com.gun0912.tedpermission.TedPermission
 import kotlinx.android.synthetic.main.fragment_oneonone_code.*
-import java.net.HttpURLConnection
 
-class CodeActivity : AppCompatActivity() {
+class JoinActivity : AppCompatActivity() {
 
     private val viewModelFactory by lazy {
-        CodeViewModelFactory(
+        JoinViewModelFactory(
             provideMatchApi(this),
             UserManageProvider(this)
         )
     }
 
-    private lateinit var viewModel: CodeViewModel
+    private lateinit var viewModel: JoinViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_code)
+        setContentView(R.layout.activity_join)
         setup()
     }
 
     private fun setup() {
         viewModel = ViewModelProviders.of(
-            this, viewModelFactory)[CodeViewModel::class.java]
+            this, viewModelFactory)[JoinViewModel::class.java]
         tv_code_title.text = "Enter Code"
         tv_code_number.text = "_______"
         iv_state.setImageResource(R.drawable.login_24_px)
@@ -70,18 +63,18 @@ class CodeActivity : AppCompatActivity() {
         val permissionListener: PermissionListener = object : PermissionListener {
             override fun onPermissionGranted() {
                 Toast.makeText(
-                    this@CodeActivity,
+                    this@JoinActivity,
                     "Permission Granted",
                     Toast.LENGTH_SHORT
                 ).show()
-                val intent = Intent(this@CodeActivity, CallActivity::class.java)
+                val intent = Intent(this@JoinActivity, CallActivity::class.java)
                 intent.putExtra("url", url)
                 startActivity(intent)
             }
 
             override fun onPermissionDenied(deniedPermissions: List<String>) {
                 Toast.makeText(
-                    this@CodeActivity,
+                    this@JoinActivity,
                     "Permission Denied\n$deniedPermissions",
                     Toast.LENGTH_SHORT
                 ).show()
