+++ app/src/main/res/layout/activity_join.xml
